Added data persistance through the setting.settings default thing for tare value and calibration factor. Started to add for export path as well but commented out. May re-add later.
Data now exports the full reports to a sub folder ("Full Results") and exports short reports to export location. (creates both if do not exist)

Added combo box which pulls suggestions for testnames from the default test report folder so adding to a given report is easier

Fixed some of the spacing issues which occur with resizing.